Fix an error with updateDinosaur not working in deleteFeature
updateDinosaur was not a function in the deleteFeature function within Dinosaur.js. I fixed this by passing in the feature index and filtering out the feature that matched that index and leaving the rest in. Also I changed the const in deleteFeature to be updatedDinosaur (which is passed into the function updateDinosaur) so it did not match the deconstructed props passed in above with the name updateDinosaur.